+++ src/service/LikePerfumeService.ts
@@ -1,5 +1,6 @@
 import LikePerfumeDao from '@src/dao/LikePerfumeDao';
 import { FailedToCreateError, NotMatchedError } from '@src/utils/errors/errors';
+import _ from 'lodash';
 
 let likePerfumeDao: LikePerfumeDao = new LikePerfumeDao();
 
@@ -29,5 +30,34 @@ export class LikePerfumeService {
             }
             throw new FailedToCreateError();
         }
+    }
+
+    async isLike(userIdx: number, perfumeIdx: number): Promise<boolean> {
+        try {
+            await this.likePerfumeDao.read(userIdx, perfumeIdx);
+            return true;
+        } catch (err) {
+            if (err instanceof NotMatchedError) {
+                return false;
+            }
+            throw err;
+        }
+    }
+
+    isLikeJob(likePerfumeList: any[]): (obj: any) => any {
+        const likeMap: { [key: string]: boolean } = _.chain(likePerfumeList)
+            .keyBy('perfumeIdx')
+            .mapValues(() => true)
+            .value();
+
+        return (obj: any) => {
+            const ret: any = Object.assign({}, obj);
+            ret.isLiked = likeMap[obj.perfumeIdx] ? true : false;
+            return ret;
+        };
+    }
+
+    async readLikeInfo(userIdx: number, perfumeIdxList: number[]) {
+        return await likePerfumeDao.readLikeInfo(userIdx, perfumeIdxList);
     }
 }
